+++ go.mod
@@ -1,0 +1,16 @@
+module github.com/hugoluchessi/go-metrics
+
+go 1.12
+
+require (
+	github.com/DataDog/datadog-go v0.0.0-20180822151419-281ae9f2d895
+	github.com/armon/go-metrics v0.0.0-20180917152333-f0300d1749da
+	github.com/circonus-labs/circonus-gometrics v2.2.6+incompatible
+	github.com/circonus-labs/circonusllhist v0.1.3 // indirect
+	github.com/hashicorp/go-immutable-radix v1.0.0
+	github.com/hashicorp/go-retryablehttp v0.5.2 // indirect
+	github.com/pascaldekloe/goe v0.0.0-20180627143212-57f6aae5913c
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/prometheus/client_golang v0.9.2
+	github.com/tv42/httpunix v0.0.0-20150427012821-b75d8614f926 // indirect
+)
