+++ providers/inmem/inmem_signal_test.go
@@ -1,4 +1,4 @@
-package metrics
+package inmem
 
 import (
 	"bytes"
@@ -7,6 +7,8 @@ import (
 	"syscall"
 	"testing"
 	"time"
+
+	"github.com/hugoluchessi/go-metrics"
 )
 
 func TestInmemSignal(t *testing.T) {
@@ -19,9 +21,9 @@ func TestInmemSignal(t *testing.T) {
 	inm.EmitKey([]string{"bar"}, 42)
 	inm.IncrCounter([]string{"baz"}, 42)
 	inm.AddSample([]string{"wow"}, 42)
-	inm.SetGaugeWithLabels([]string{"asdf"}, 42, []Label{{"a", "b"}})
-	inm.IncrCounterWithLabels([]string{"qwer"}, 42, []Label{{"a", "b"}})
-	inm.AddSampleWithLabels([]string{"zxcv"}, 42, []Label{{"a", "b"}})
+	inm.SetGaugeWithLabels([]string{"asdf"}, 42, []metrics.Label{{"a", "b"}})
+	inm.IncrCounterWithLabels([]string{"qwer"}, 42, []metrics.Label{{"a", "b"}})
+	inm.AddSampleWithLabels([]string{"zxcv"}, 42, []metrics.Label{{"a", "b"}})
 
 	// Wait for period to end
 	time.Sleep(15 * time.Millisecond)
