+++ providers/statsite/statsite_test.go
@@ -1,12 +1,12 @@
-package metrics
+package statsite
 
 import (
 	"bufio"
 	"net"
-	"net/url"
-	"strings"
 	"testing"
 	"time"
+
+	"github.com/hugoluchessi/go-metrics"
 )
 
 func acceptConn(addr string) net.Conn {
@@ -108,64 +108,26 @@ func TestStatsite_Conn(t *testing.T) {
 		conn.Close()
 		done <- true
 	}()
+
+	time.Sleep(1 * time.Millisecond)
+
 	s, err := NewStatsiteSink(addr)
 	if err != nil {
 		t.Fatalf("bad error")
 	}
 
 	s.SetGauge([]string{"gauge", "val"}, float32(1))
-	s.SetGaugeWithLabels([]string{"gauge_labels", "val"}, float32(2), []Label{{"a", "label"}})
+	s.SetGaugeWithLabels([]string{"gauge_labels", "val"}, float32(2), []metrics.Label{{"a", "label"}})
 	s.EmitKey([]string{"key", "other"}, float32(3))
 	s.IncrCounter([]string{"counter", "me"}, float32(4))
-	s.IncrCounterWithLabels([]string{"counter_labels", "me"}, float32(5), []Label{{"a", "label"}})
+	s.IncrCounterWithLabels([]string{"counter_labels", "me"}, float32(5), []metrics.Label{{"a", "label"}})
 	s.AddSample([]string{"sample", "slow thingy"}, float32(6))
-	s.AddSampleWithLabels([]string{"sample_labels", "slow thingy"}, float32(7), []Label{{"a", "label"}})
+	s.AddSampleWithLabels([]string{"sample_labels", "slow thingy"}, float32(7), []metrics.Label{{"a", "label"}})
 
 	select {
 	case <-done:
 		s.Shutdown()
 	case <-time.After(3 * time.Second):
 		t.Fatalf("timeout")
-	}
-}
-
-func TestNewStatsiteSinkFromURL(t *testing.T) {
-	for _, tc := range []struct {
-		desc       string
-		input      string
-		expectErr  string
-		expectAddr string
-	}{
-		{
-			desc:       "address is populated",
-			input:      "statsd://statsd.service.consul",
-			expectAddr: "statsd.service.consul",
-		},
-		{
-			desc:       "address includes port",
-			input:      "statsd://statsd.service.consul:1234",
-			expectAddr: "statsd.service.consul:1234",
-		},
-	} {
-		t.Run(tc.desc, func(t *testing.T) {
-			u, err := url.Parse(tc.input)
-			if err != nil {
-				t.Fatalf("error parsing URL: %s", err)
-			}
-			ms, err := NewStatsiteSinkFromURL(u)
-			if tc.expectErr != "" {
-				if !strings.Contains(err.Error(), tc.expectErr) {
-					t.Fatalf("expected err: %q, to contain: %q", err, tc.expectErr)
-				}
-			} else {
-				if err != nil {
-					t.Fatalf("unexpected err: %s", err)
-				}
-				is := ms.(*StatsiteSink)
-				if is.addr != tc.expectAddr {
-					t.Fatalf("expected addr %s, got: %s", tc.expectAddr, is.addr)
-				}
-			}
-		})
 	}
 }
